feat: solve main task of Worker Threads
