+++ src/wt/worker.js
@@ -1,36 +1,13 @@
-import path from 'path';
-import { argv } from 'process';
-import { fileURLToPath } from 'node:url';
-import { Worker, isMainThread, parentPort, workerData } from 'node:worker_threads';
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const pathToWorker = path.join(__dirname, 'worker.js');
+import { workerData, parentPort } from 'worker_threads';
 
-const num = argv[2];
-
+// n should be received from main thread
 export const nthFibonacci = (n) => n < 2 ? n : nthFibonacci(n - 1) + nthFibonacci(n - 2);
 
 export const sendResult = () => {
-  if (isMainThread) {
-    return new Promise((resolve, reject) => {
-      const worker = new Worker(pathToWorker, {
-        workerData: String(nthFibonacci)
-      });
-      worker.postMessage(num);
-      worker.on('message', msg => console.log('Worker send message:', msg));
-      worker.on('message', resolve);
-      worker.on('error', reject);
-      worker.on('exit', (code) => {
-        if (code !== 0)
-          reject(new Error(`Worker stopped with exit code ${code}`));
-      });
-    });
-  } else {
-    parentPort.once('message', (message = 7) => {
-      const result = eval(workerData)(message);
-      parentPort.postMessage(result);
-    })
-  }
+    // This function sends result of nthFibonacci computations to main thread
+    const result = nthFibonacci(workerData);
+    parentPort.postMessage(result);
 };
 
 sendResult();
